added ui stuff

+++ src/main/java/com/stefanmocoat/showoffice/jpa/repository/PferdRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.stefanmocoat.showoffice.jpa.entities.Pferd;
 
 @Repository
-public interface PferdRepository extends JpaRepository<Pferd, Long> {
+public interface PferdRepository extends JpaRepository<Pferd, Integer> {
 
     Pferd findByKopfnummer(String kopfnummer);
